#5: move IdleTime to Example project
