+++ Example/2005-Example.csproj
@@ -125,6 +125,7 @@
     <Compile Include="ConferenceForm.cs">
       <SubType>Form</SubType>
     </Compile>
+    <Compile Include="IdleTime.cs" />
     <Compile Include="MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
