+++ jabber-net.csproj
@@ -192,7 +192,6 @@
     <Compile Include="bedrock\util\GetOptBase.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="bedrock\util\IdleTime.cs" />
     <Compile Include="bedrock\util\Version.cs">
       <SubType>Code</SubType>
     </Compile>
